+++ gitbook2pdf/__init__.py
@@ -1,2 +1,3 @@
 from .gitbook2pdf import Gitbook2PDF
+
 __all__ = ('Gitbook2PDF',)
